requirements.txt
scikit-learn~=1.0.2
pandas~=1.4.1
fastapi~=0.85.0
uvicorn~=0.18.3
pytest~=7.1.2
httpx~=0.19.0
loguru~=0.5.3
celery~=5.2.7
redis~=3.5.3
python-dotenv~=0.21.0
requests